fastuidraw/painter: move StrokedPath::point to global class StrokedPoint
