+++ src/fastuidraw/painter/Rules.mk
@@ -20,7 +20,7 @@ FASTUIDRAW_SOURCES += $(call filelist, fill_rule.cpp \
 	painter_shader.cpp painter_shader_set.cpp \
 	painter_dashed_stroke_shader_set.cpp painter_stroke_shader.cpp \
 	painter_glyph_shader.cpp painter_blend_shader_set.cpp \
-	painter_fill_shader.cpp \
+	painter_fill_shader.cpp stroked_point.cpp \
 	stroked_path.cpp filled_path.cpp)
 
 # Begin standard footer
